+++ examples/python_rs/llm/vllm/disaggregated/prefill_worker.py
@@ -18,12 +18,11 @@ import asyncio
 
 import uvloop
 import vllm
+from common.parser import parse_vllm_args
+from common.protocol import PrefillRequest, PrefillResponse
 from triton_distributed_rs import DistributedRuntime, triton_endpoint, triton_worker
 from vllm.engine.arg_utils import AsyncEngineArgs
 from vllm.logger import logger as vllm_logger
-
-from .parser import parse_vllm_args
-from .protocol import PrefillRequest, PrefillResponse
 
 
 class VllmPrefillEngine:
